3.1.5

+++ route/edit_delete.py
@@ -6,33 +6,36 @@ def edit_delete_2(conn, name, app_var):
     ip = ip_check()
     if acl_check(name) == 1:
         return re_error('/ban')
-    
+
     if flask.request.method == 'POST':
         if captcha_post(flask.request.form.get('g-recaptcha-response', '')) == 1:
             return re_error('/error/13')
         else:
             captcha_post('', 0)
 
+        if slow_edit_check() == 1:
+            return re_error('/error/24')
+
         curs.execute(db_change("select data from data where title = ?"), [name])
         data = curs.fetchall()
         if data:
             today = get_time()
             leng = '-' + str(len(data[0][0]))
-            
+
             history_plus(
-                name, 
-                '', 
-                today, 
-                ip, 
-                flask.request.form.get('send', ''), 
+                name,
+                '',
+                today,
+                ip,
+                flask.request.form.get('send', ''),
                 leng,
                 'delete'
             )
-            
+
             curs.execute(db_change("select title, link from back where title = ? and not type = 'cat' and not type = 'no'"), [name])
             for data in curs.fetchall():
                 curs.execute(db_change("insert into back (title, link, type) values (?, ?, 'no')"), [data[0], data[1]])
-            
+
             curs.execute(db_change("delete from back where link = ?"), [name])
             curs.execute(db_change("delete from data where title = ?"), [name])
             conn.commit()
@@ -42,16 +45,19 @@ def edit_delete_2(conn, name, app_var):
             file_check = file_check.groups()
             os.remove(os.path.join(
                 app_var['path_data_image'],
-                hashlib.sha224(bytes(file_check[0], 'utf-8')).hexdigest() + '.' + file_check[1]
+                sha224_replace(file_check[0], 'utf-8') + '.' + file_check[1]
             ))
-            
+
+        curs.execute(db_change('select data from other where name = "count_all_title"'))
+        curs.execute(db_change("update other set data = ? where name = 'count_all_title'"), [str(int(curs.fetchall()[0][0]) - 1)])
+
         return redirect('/w/' + url_pas(name))
     else:
         curs.execute(db_change("select title from data where title = ?"), [name])
         if not curs.fetchall():
             return redirect('/w/' + url_pas(name))
 
-        return easy_minify(flask.render_template(skin_check(), 
+        return easy_minify(flask.render_template(skin_check(),
             imp = [name, wiki_set(), custom(), other2([' (' + load_lang('delete') + ')', 0])],
             data =  '''
                     <form method="post">
